Move ToJSON and FastToJSON to outputnode.go
